+++ jax/_src/stages.py
@@ -33,6 +33,7 @@ from __future__ import annotations
 from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import Any, NamedTuple, Protocol, Union
+import warnings
 
 import jax
 
@@ -318,6 +319,11 @@ class XlaLowering(Lowering):
 
   def mhlo(self) -> ir.Module:
     """Return an MHLO representation of this computation."""
+    warnings.warn(
+        "mhlo support is deprecated and will be removed "
+        "from a future release of JAX. Use stablehlo instead.",
+        DeprecationWarning,
+    )
     module_str = xla_extension.mlir.stablehlo_to_mhlo(
         mlir.module_to_bytecode(self.stablehlo()))
     with self.stablehlo().context: